suseconds_t is a long on Linux

+++ include/kernel/linux/sys/resource.h
@@ -27,7 +27,7 @@ typedef long long rlim_t;
 
 # ifndef suseconds_t
 #  define suseconds_t suseconds_t
-typedef int suseconds_t;
+typedef long suseconds_t;
 # endif
 # ifndef time_t
 #  define time_t time_t
