+++ include/kernel/linux/sys/time.h
@@ -22,7 +22,7 @@
 /* types */
 # ifndef suseconds_t
 #  define suseconds_t suseconds_t
-typedef int suseconds_t;
+typedef long suseconds_t;
 # endif
 # ifndef time_t
 #  define time_t time_t
