+++ include/kernel/linux/sys/types.h
@@ -50,7 +50,7 @@ typedef long long off_t;
 # endif
 # ifndef suseconds_t
 #  define suseconds_t suseconds_t
-typedef int suseconds_t;
+typedef long suseconds_t;
 # endif
 # ifndef time_t
 #  define time_t time_t
